improve Sona reliability and organization
