+++ .gitignore
@@ -37,7 +37,7 @@ ISSUES-REPORT.md
 desktop/src-tauri/binaries/
 !desktop/src-tauri/binaries/.gitkeep
 *.csv
-sona/
+/sona/
 whisper.cpp/
 aptabase-python/
 tauri-plugin-aptabase/
