+++ .sona-version
@@ -1,1 +1,1 @@
-v0.2.2
+v0.2.3
