+++ desktop/src-tauri/Cargo.toml
@@ -36,7 +36,8 @@ serde_json = { workspace = true }
 eyre = { workspace = true }
 
 tokio = { version = "1.49.0", features = ["net", "fs", "macros", "rt-multi-thread"] }
-tokio-util = { version = "0.7", features = ["io"] }
+tokio-util = { version = "0.7", features = ["codec", "io"] }
+bytes = "1"
 once_cell = "1.21.3"
 serde = { version = "^1.0.228", features = ["derive"] }
 url = "2.5.8"
