+++ desktop/src-tauri/src/cmd/sona_cmd.rs
@@ -74,6 +74,19 @@ pub async fn load_model(app_handle: tauri::AppHandle, model_path: String, gpu_de
     let sona_state: State<'_, Mutex<SonaState>> = app_handle.state();
     let mut state_guard = sona_state.lock().await;
 
+    let process_is_alive = state_guard
+        .process
+        .as_mut()
+        .is_some_and(crate::sona::SonaProcess::is_alive);
+    if !process_is_alive {
+        if state_guard.process.is_some() {
+            tracing::warn!("cached sona process is no longer running; restarting it");
+        }
+        state_guard.process = None;
+        state_guard.loaded_model_path = None;
+        state_guard.loaded_gpu_device = None;
+    }
+
     // Check if model already loaded with same gpu_device
     if let Some(ref loaded_path) = state_guard.loaded_model_path {
         if *loaded_path == model_path && state_guard.loaded_gpu_device == gpu_device {
